Make changes to files
Adjust the forEach method to create a squareContainer variable that holds the containers for squares. Create a for loop within the method that will determine if the initializer variable holds an integer that can be divided by 16 without a remainder. If so, then the squareContainer will be appended to the grid with a class attached to it as well.

Attempt at nesting a for loop within the loop as well.

Create .squareContainer class in style.css and set it's display property to flex